requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic[email]==2.10.6
psutil==7.0.0
python-multipart==0.0.20
